chore(controller): run without k8s cluster
