+++ server/controller/src/main/java/ai/starwhale/mlops/api/JobController.java
@@ -46,6 +46,7 @@ import ai.starwhale.mlops.exception.SwValidationException.ValidSubject;
 import ai.starwhale.mlops.exception.api.StarwhaleApiException;
 import com.github.pagehelper.PageInfo;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -54,6 +55,7 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 @RestController
 @RequestMapping("${sw.controller.api-prefix}")
+@ConditionalOnProperty(name = "sw.modules.jobs.enabled", havingValue = "true")
 public class JobController implements JobApi {
 
     private final JobService jobService;
