+++ server/controller/src/main/java/ai/starwhale/mlops/domain/job/ModelServingService.java
@@ -70,12 +70,14 @@ import javax.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.http.HttpStatus;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 @Slf4j
 @Service
+@ConditionalOnProperty(value = "sw.modules.jobs.enabled", havingValue = "true")
 public class ModelServingService {
 
     private final ModelServingMapper modelServingMapper;
