+++ server/controller/src/main/resources/application.yaml
@@ -14,6 +14,9 @@ sw:
   version: ${SW_VERSION_CONTROLLER:0.1.0:8c82767b60686f3e2bfea9dafe8c8cce5dd34f52}
   features:
     disabled: ${SW_DISABLED_FEATURES:}
+  modules:
+    jobs:
+      enabled: ${SW_MODULES_JOBS_ENABLED:true}
   public-api:
     base-url: ${SW_PUBLIC_API_BASE_URL:}
     latest-version: ${SW_LATEST_VERSION_URL:${sw.public-api.base-url}/version/latest}
@@ -44,6 +47,7 @@ sw:
   instance-uri: ${SW_INSTANCE_URI:http://console.pre.intra.starwhale.ai}
   infra:
     k8s:
+      enabled: ${SW_INFRA_K8S_ENABLED:true}
       name-space: ${SW_K8S_NAME_SPACE:default}
       host-path-for-cache: ${SW_K8S_HOST_PATH_FOR_CACHE:}
       model-serving-template-path: ${SW_K8S_MODEL_SERVING_TEMPLATE_PATH:}
